requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
executing==2.0.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
future==1.0.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
line-bot-sdk==3.11.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.0.5
openai==1.31.1
packaging==24.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.46
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
uWSGI==2.0.26
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
